w8d3 code update

+++ xpsc/week8/day3/NASA.cpp
@@ -53,7 +53,7 @@ int main()
             count[a[i]]++;
         }
 
-        long long ans = n;
+        ll ans = n;
         for (int i = 0; i < n; i++)
         {
             for (int j = 0; j < allPalindrome.size(); j++)
